Fix authentication state when calling syncIdentifiers (#649)
* Set authentication state as object instead of raw value in syncIdentifiers call.
Send authentication state as MobileVisitorAuthenticationState instead of Int in event data for syncIdentifiers API call. The Identity extension casts the authentication state as a MobileVisitorAuthenticationState when extracting from the event data. When using the raw value (Int) of the MobileVisitorAuthenticationState enum, the cast always fails and returns value .unknown. This fix will allow the user specified authentication state for the custom identifier in the syncIdentifiers call instead of all authentication states being unknown.

* Verify authentication state as object and not int in Identity unit tests